Move GBN to its own file
